# Trivial optimize of function coalesce. 