--- a/src/Functions/coalesce.cpp
+++ b/src/Functions/coalesce.cpp
@@ -33,6 +33,8 @@ public:
         : context(context_)
         , is_not_null(FunctionFactory::instance().get("isNotNull", context))
         , assume_not_null(FunctionFactory::instance().get("assumeNotNull", context))
+        , if_function(FunctionFactory::instance().get("if", context))
+        , multi_if_function(FunctionFactory::instance().get("multiIf", context))
     {
     }
 
@@ -150,13 +152,8 @@ public:
         /// If there was only two arguments (3 arguments passed to multiIf)
         /// use function "if" instead, because it's implemented more efficient.
         /// TODO: make "multiIf" the same efficient.
-        FunctionOverloadResolverPtr if_function;
-        if (multi_if_args.size() == 3)
-            if_function = FunctionFactory::instance().get("if", context);
-        else
-            if_function = FunctionFactory::instance().get("multiIf", context);
-
-        ColumnPtr res = if_function->build(multi_if_args)->execute(multi_if_args, result_type, input_rows_count);
+        FunctionOverloadResolverPtr if_or_multi_if = multi_if_args.size() == 3 ? if_function : multi_if_function;
+        ColumnPtr res = if_or_multi_if->build(multi_if_args)->execute(multi_if_args, result_type, input_rows_count);
 
         /// if last argument is not nullable, result should be also not nullable
         if (!multi_if_args.back().column->isNullable() && res->isNullable())
@@ -176,6 +173,8 @@ private:
     ContextPtr context;
     FunctionOverloadResolverPtr is_not_null;
     FunctionOverloadResolverPtr assume_not_null;
+    FunctionOverloadResolverPtr if_function;
+    FunctionOverloadResolverPtr multi_if_function;
 };
 
 }