--- a/tests/performance/coalesce.xml
+++ b/tests/performance/coalesce.xml
@@ -1,0 +1,3 @@
+<test>
+    <query>select coalesce(materialize(null), -1) from numbers(1000000000) format Null settings max_block_size = 8192</query>
+</test>