### Changelog category (leave one):
- Performance Improvement

### Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
Reuse the result of `FunctionFactory::instance().get("isNotNull", context)` and `FunctionFactory::instance().get("assumeNotNull", context)`. Make sure it is called once during the lifetime of `FunctionCoalesce`. 


Notice that the cost of `FunctionFactory::instance().get` can't be ignored during the execution of function `coalesce`. In my tests: 
- `executeImpl` costs 68068 ns
- `FunctionFactory::instance().get` inside `executeImpl` costs 9537 ns, which is 14% of `executeImpl`

``` sql 
select coalesce(materialize(null), -1) from numbers(1000000000) format Null settings max_block_size = 8192;  
Old: 0 rows in set. Elapsed: 2.478 sec. Processed 1.00 billion rows, 8.00 GB (403.61 million rows/s., 3.23 GB/s.)
New: 0 rows in set. Elapsed: 2.321 sec. Processed 1.00 billion rows, 8.00 GB (430.86 million rows/s., 3.45 GB/s.)
```